feat: stream_nft_exists view Ix
test: streamNftExists

add note comment

refactor: rename to stream_exists

polish comment

test: add view dir

remove note in stream_exists
