+++ programs/lockup/src/instructions/view/mod.rs
@@ -1,5 +1,6 @@
 pub mod refundable_amount_of;
 pub mod status_of;
+pub mod stream_exists;
 pub mod stream_view;
 pub mod streamed_amount_of;
 pub mod treasury_view;
@@ -7,6 +8,7 @@ pub mod withdrawable_amount_of;
 pub mod withdrawal_fee_in_lamports;
 
 pub use status_of::*;
+pub use stream_exists::*;
 pub use stream_view::*;
 pub use treasury_view::*;
 pub use withdrawal_fee_in_lamports::*;
