+++ tests/lockup/unit/view/streamedAmountOf.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, it } from "vitest";
-import { ZERO } from "../../../lib/constants";
-import { toBn } from "../../../lib/helpers";
-import { assertEqBn } from "../../common/assertions";
-import { LockupTestContext } from "../context";
-import { getStreamedAmount } from "../utils/calculations";
-import { AMOUNTS, Amount, TIMESTAMPS, Time, UNLOCK_AMOUNTS } from "../utils/defaults";
+import { ZERO } from "../../../../lib/constants";
+import { toBn } from "../../../../lib/helpers";
+import { assertEqBn } from "../../../common/assertions";
+import { LockupTestContext } from "../../context";
+import { getStreamedAmount } from "../../utils/calculations";
+import { AMOUNTS, Amount, TIMESTAMPS, Time, UNLOCK_AMOUNTS } from "../../utils/defaults";
 
 let ctx: LockupTestContext;
