+++ tests/lockup/unit/view/treasuryView.test.ts
@@ -1,30 +1,30 @@
 import { PublicKey } from "@solana/web3.js";
 import { beforeEach, describe, expect, it } from "vitest";
-import { ProgramId } from "../../../lib/constants";
-import { assertEqTreasury } from "../../common/assertions";
-import { MerkleInstantTestContext } from "../context";
-import { Seed } from "../utils/defaults";
+import { ProgramId } from "../../../../lib/constants";
+import { assertEqTreasury } from "../../../common/assertions";
+import { LockupTestContext } from "../../context";
+import { Seed } from "../../utils/defaults";
 
 describe("treasuryView", () => {
-  let ctx: MerkleInstantTestContext;
+  let ctx: LockupTestContext;
 
   describe("given not initialized", () => {
     it("should fail", async () => {
-      ctx = new MerkleInstantTestContext();
-      await ctx.setUpMerkleInstant({ initProgram: false });
+      ctx = new LockupTestContext();
+      await ctx.setUpLockup({ initProgram: false });
       await expect(ctx.treasuryView()).rejects.toThrow("Cannot read properties of null (reading 'data')");
     });
   });
 
   describe("given initialized", () => {
     beforeEach(async () => {
-      ctx = new MerkleInstantTestContext();
-      await ctx.setUpMerkleInstant();
+      ctx = new LockupTestContext();
+      await ctx.setUpLockup();
     });
 
     it("should return the treasury details", async () => {
       const actualTreasury = await ctx.treasuryView();
-      const expectedBump = PublicKey.findProgramAddressSync([Seed.TREASURY], ctx.merkleInstant.programId)[1];
+      const expectedBump = PublicKey.findProgramAddressSync([Seed.TREASURY], ctx.lockup.programId)[1];
       const expectedTreasury = {
         bump: expectedBump,
         chainlinkProgram: ProgramId.CHAINLINK_PROGRAM,
